// projects/chantiers360-v2/app/lib/planning/types.ts
export type SimplePlanningInput = {
  plannedStartDate?: string | null;
  plannedEndDate?: string | null;
  nextInterventionDate?: string | null;
  upcomingMilestone?: string | null;
  simplePlanningComment?: string | null;
};

export type SimplePlanningData = {
  plannedStartDate: string | null;
  plannedEndDate: string | null;
  nextInterventionDate: string | null;
  upcomingMilestone: string | null;
  simplePlanningComment: string | null;
};